renamed route, validator homeworksList type prop, template v-if
